Create index.html
Task #3
Create a simple web page that allows the user to encrypt and decrypt text using the Web Crypto API. The page should have two text boxes, one for the plaintext and one for the ciphertext. The user should be able to enter the plaintext in the first text box and then click on a button to encrypt the text. The encrypted text should then be displayed in the second text box. The user should also be able to enter the ciphertext in the second text box and then click on a button to decrypt the text. The decrypted text should then be displayed in the first text box.